added php changes

+++ js/Detailview.php
@@ -1,35 +1,56 @@
 <?php
 $iti = new RecursiveDirectoryIterator(getcwd() . '/resources/XMLVarad');
-$numberOfJsonFiles = 0;
-$currentFileNumber = 0;
-foreach (new RecursiveIteratorIterator($iti) as $file) {
+$fileArray = array();
+$recursiveIti = new RecursiveIteratorIterator($iti);
+$workDir = str_replace('\\',"/",getcwd());
+$fp = fopen($workDir . '/resources/XMLVarad/varamuteKogumik.json', 'a');
+
+if(!file_exists($workDir . '/resources/XMLVarad/varamuteKogumik.json')){
+    header("Location: 404.html");
+    exit();
+}
+
+$fileContent = file_get_contents($workDir . "/resources/XMLVarad/varamuteKogumik.json");
+
+
+foreach ($recursiveIti as $file) {
     if (strpos($file, 'varamu.json') !== false) {
-        $numberOfJsonFiles++;
+        array_push($fileArray,$file);
+        @$numberOfJsonFiles++;
     }
 }
-$fp = fopen(getcwd() . '/resources/XMLVarad/varamuteKogumik.json', 'a');
-file_put_contents(getcwd() . "/resources/XMLVarad/varamuteKogumik.json", "");
-fwrite($fp, '[');
-foreach (new RecursiveIteratorIterator($iti) as $file) {
+
+@$resourceString .= '[';
+
+foreach ($fileArray as $file){
     if (strpos($file, 'varamu.json') !== false) {
         $string = file_get_contents($file);
         $json_a = json_decode($string, true);
-
-        foreach ($json_a as $person_a) {
-            $currentFileNumber++;
+        foreach ($json_a as $asset) {
+            @$currentFileNumber++;
             if ($currentFileNumber == $numberOfJsonFiles) {
-                fwrite($fp, json_encode($person_a));
+                $resourceString .= json_encode($asset);
             } else {
-                fwrite($fp, json_encode($person_a) . ',');
+                $resourceString .= json_encode($asset) . ',';
             }
         }
     }
 }
-fwrite($fp, ']');
-fclose($fp);
-?>
-<script type='text/javascript'>
 
+$resourceString .= ']';
+
+writeToFile($fileContent, $resourceString, $fp);
+
+function writeToFile($existingContent, $string, $resourcePath){
+    if($existingContent != $string){
+        file_put_contents(getcwd() . "/resources/XMLVarad/varamuteKogumik.json", "");
+        fwrite($resourcePath, $string);
+        fclose($resourcePath);
+    }
+}
+?>
+
+<script type='text/javascript'>
 
     if(!(location.hash==='')){
         console.log(location.hash);
@@ -38,26 +59,21 @@ fclose($fp);
         $(document).ready(function () {
             var filesArrayPHP = [];
             <?php
-            $it = new RecursiveDirectoryIterator(getcwd() . "/resources/XMLVarad");
-            foreach (new RecursiveIteratorIterator($it) as $file) {
+            foreach ($recursiveIti as $file) {
                 if (substr($file, -1) != ".." && substr($file, -1) != "." && strpos($file, '.DS_Store') == false) {
-                    echo 'filesArrayPHP.push("' . $file . '");';
+                    $filePath = substr($file, strpos($file, '/resources'));
+                    $correctFilePath = str_replace('\\',"/",$filePath);
+                    $correctEncoding = mb_convert_encoding($correctFilePath, 'UTF-8', mb_detect_encoding($correctFilePath));
+                    echo 'filesArrayPHP.push("' . $correctEncoding . '");';
                 }
             }
-
-
             ?>
-
-
             $('.table-responsive').hide();
             var template = $('#main-row-template').html();
             var id = "";
             var url = 'https://' + window.location.hostname;
-
-
             $.getJSON('resources/XMLVarad/varamuteKogumik.json', function (data) {
                 var usedNames = [];
-
                 var resource = data[pageId];
                 id = resource;
                 var temp = $('#detail-row-template').html();
@@ -74,11 +90,9 @@ fclose($fp);
                         }
                     });
                 }
-
                 $('.versionButton').click(function (event) {
                     $('.versionButton').attr("class", "btn btn-outline-primary btn-sm versionButton");
                     $(this).attr("class", "btn btn-primary btn-sm versionButton");
-
                     resource = data[event.target.id];
                     id = resource;
                     newRow.find('#name_est').text(resource.Name_est);
@@ -92,10 +106,8 @@ fclose($fp);
                     newRow.find('#resource_type').text(resource.Resource_type);
                     newRow.find('#owner').text(resource.Owner);
                     newRow.find('#field').text(resource.Field);
-
                     dtbody.append(newRow);
                 });
-
                 newRow.find('#name_est').text(resource.Name_est);
                 newRow.find('#name_eng').text(resource.Name_eng);
                 newRow.find('#version').text(resource.Version);
@@ -105,15 +117,11 @@ fclose($fp);
                 newRow.find('#des_eng').text(resource.Description_eng);
                 newRow.find('#resource_type').text(resource.Resource_type);
                 newRow.find('#owner').text(resource.Owner);
-
                 dtbody.append(newRow);
-
                 $('.table-responsive').hide();
                 $('#detail').show().css("background-color", "white");
                 $('#files').hide();
-
                 $('#detail-tbody th').css('white-space', 'nowrap');
-
 
                 $('.fileButton').on('click', function (event) {
                     $('#files-content').empty();
@@ -126,23 +134,21 @@ fclose($fp);
                     var directoryVersion = id.Version;
                     $('#files-content').append("<h3>" + fileName + "</h3>");
                     for (var index = 0; index < filesArrayPHP.length; index++) {
-                        var filePath = replaceAllSpecialChars(filesArrayPHP[index]);
-                        var adjustedURL = reworkFilePath(filePath);
+                        var filePath = filesArrayPHP[index]
                         if (filePath.indexOf(directoryName) >= 0 && filePath.indexOf(directoryVersion) >= 0 && !filePath.includes('varamu.json')) {
-                            $('#files-content').append('<a href="' + '.' + adjustedURL + '" download>' + findFileName(filePath, directoryName + '_V' + directoryVersion) + "</a><br>");
+                            $('#files-content').append('<a href="' + '.' + filePath + '" download>' + findFileName(filePath, directoryName + '_V' + directoryVersion) + "</a><br>");
                             $('#files-content').append(
-                                '<details><summary>URL</summary><p>' + url + adjustedURL.replace('.', '') + '</p></details><br>'
+                                '<details><summary>URL</summary><p>' + url + filePath.replace('.', '') + '</p></details><br>'
                             );
                         } else if (filePath.indexOf(directoryName) >= 0 && (directoryVersion.length > 3 || directoryVersion.length == 0) && !filePath.includes('varamu.json')) {
-                            $('#files-content').append('<a href="' + '.' + adjustedURL + '" download>' + findFileName(filePath, directoryName) + "</a><br>");
+                            $('#files-content').append('<a href="' + '.' + filePath + '" download>' + findFileName(filePath, directoryName) + "</a><br>");
                             $('#files-content').append(
-                                '<details><summary>URL</summary><p>' + url + adjustedURL.replace('.', '') + '</p></details><br>'
+                                '<details><summary>URL</summary><p>' + url + filePath.replace('.', '') + '</p></details><br>'
                             );
                         }
                     }
                     $('summary').css('cursor', 'pointer');
                 });
-
                 $('.name_and_version').css("color", "#023cc1").hover(function () {
                     $(this).css("cursor", 'pointer');
                     $(this).css("color", "black");
@@ -150,7 +156,6 @@ fclose($fp);
                     $(this).css("color", "#023cc1");
                 });
             });
-
             $('.backButton').on('click', function (event) {
                 window.location = '/index.php';
             });
@@ -161,37 +166,29 @@ fclose($fp);
                 $('.versionButton').show();
                 $('#detail').show();
             });
-
-
         });
-
     }else {
         $(document).ready(function () {
             var filesArrayPHP = [];
             <?php
-            $it = new RecursiveDirectoryIterator(getcwd() . "/resources/XMLVarad");
-            foreach (new RecursiveIteratorIterator($it) as $file) {
+            foreach ($recursiveIti as $file) {
                 if (substr($file, -1) != ".." && substr($file, -1) != "." && strpos($file, '.DS_Store') == false) {
-                    echo 'filesArrayPHP.push("' . $file . '");';
+                    $filePath = substr($file, strpos($file, '/resources'));
+                    $correctFilePath = str_replace('\\',"/",$filePath);
+                    $correctEncoding = mb_convert_encoding($correctFilePath, 'UTF-8', mb_detect_encoding($correctFilePath));
+                    echo 'filesArrayPHP.push("' . $correctEncoding . '");';
                 }
             }
-
-
             ?>
-
-
             $('#detail').hide();
             var template = $('#main-row-template').html();
             var tBody = $('#xml-resources');
             var id = "";
             var url = 'https://' + window.location.hostname;
-
-
             $.getJSON('resources/XMLVarad/varamuteKogumik.json', function (data) {
                 var resourceName = "";
                 var resourceVersion = "";
                 var usedNames = [];
-
                 //show only repeating names' latest version
                 data.forEach(function (resource, index) {
                     var newRow = $(template);
@@ -210,7 +207,6 @@ fclose($fp);
                     }
                     $('#files').hide();
                 });
-
                 //go through all again, but show only non-duplicates
                 data.forEach(function (resource, index) {
                     var newRow = $(template);
@@ -221,8 +217,6 @@ fclose($fp);
                         tBody.append(newRow);
                     }
                 });
-
-
                 $('.name_and_version').click(function (event) {
                     var resource = data[event.target.id];
                     id = resource;
@@ -240,11 +234,9 @@ fclose($fp);
                             }
                         });
                     }
-
                     $('.versionButton').click(function (event) {
                         $('.versionButton').attr("class", "btn btn-outline-primary btn-sm versionButton");
                         $(this).attr("class", "btn btn-primary btn-sm versionButton");
-
                         resource = data[event.target.id];
                         id = resource;
                         newRow.find('#name_est').text(resource.Name_est);
@@ -258,10 +250,8 @@ fclose($fp);
                         newRow.find('#resource_type').text(resource.Resource_type);
                         newRow.find('#owner').text(resource.Owner);
                         newRow.find('#field').text(resource.Field);
-
                         dtbody.append(newRow);
                     });
-
                     newRow.find('#name_est').text(resource.Name_est);
                     newRow.find('#name_eng').text(resource.Name_eng);
                     newRow.find('#version').text(resource.Version);
@@ -271,16 +261,12 @@ fclose($fp);
                     newRow.find('#des_eng').text(resource.Description_eng);
                     newRow.find('#resource_type').text(resource.Resource_type);
                     newRow.find('#owner').text(resource.Owner);
-
                     dtbody.append(newRow);
-
                     $('.table-responsive').hide();
                     $('#detail').show().css("background-color", "white");
                     $('#files').hide();
-
                     $('#detail-tbody th').css('white-space', 'nowrap');
                 });
-
                 $('.fileButton').on('click', function (event) {
                     $('#files-content').empty();
                     $('#detail').hide();
@@ -292,23 +278,21 @@ fclose($fp);
                     var directoryVersion = id.Version;
                     $('#files-content').append("<h3>" + fileName + "</h3>");
                     for (var index = 0; index < filesArrayPHP.length; index++) {
-                        var filePath = replaceAllSpecialChars(filesArrayPHP[index]);
-                        var adjustedURL = reworkFilePath(filePath);
+                        var filePath = filesArrayPHP[index]
                         if (filePath.indexOf(directoryName) >= 0 && filePath.indexOf(directoryVersion) >= 0 && !filePath.includes('varamu.json')) {
-                            $('#files-content').append('<a href="' + '.' + adjustedURL + '" download>' + findFileName(filePath, directoryName + '_V' + directoryVersion) + "</a><br>");
+                            $('#files-content').append('<a href="' + '.' + filePath + '" download>' + findFileName(filePath, directoryName + '_V' + directoryVersion) + "</a><br>");
                             $('#files-content').append(
-                                '<details><summary>URL</summary><p>' + url + adjustedURL.replace('.', '') + '</p></details><br>'
+                                '<details><summary>URL</summary><p>' + url + filePath.replace('.', '') + '</p></details><br>'
                             );
                         } else if (filePath.indexOf(directoryName) >= 0 && (directoryVersion.length > 3 || directoryVersion.length == 0) && !filePath.includes('varamu.json')) {
-                            $('#files-content').append('<a href="' + '.' + adjustedURL + '" download>' + findFileName(filePath, directoryName) + "</a><br>");
+                            $('#files-content').append('<a href="' + '.' + filePath + '" download>' + findFileName(filePath, directoryName) + "</a><br>");
                             $('#files-content').append(
-                                '<details><summary>URL</summary><p>' + url + adjustedURL.replace('.', '') + '</p></details><br>'
+                                '<details><summary>URL</summary><p>' + url + filePath.replace('.', '') + '</p></details><br>'
                             );
                         }
                     }
                     $('summary').css('cursor', 'pointer');
                 });
-
                 $('.name_and_version').css("color", "#023cc1").hover(function () {
                     $(this).css("cursor", 'pointer');
                     $(this).css("color", "black");
@@ -316,7 +300,6 @@ fclose($fp);
                     $(this).css("color", "#023cc1");
                 });
             });
-
             $('.backButton').on('click', function (event) {
                 window.location = '/index.php';
             });
@@ -327,26 +310,11 @@ fclose($fp);
                 $('.versionButton').show();
                 $('#detail').show();
             });
-
-
         });
     }
 
 
-    function replaceAllSpecialChars(string) {
-        return string;
-        result = string.replace(/Õ/g, "Õ");
-        result2 = result.replace(/ä/g, "ä");
-        result3 = result2.replace(/Ü/g, "Ü");
-        result4 = result3.replace(/õ/g, "õ");
-        return result4;
-    }
-    function reworkFilePath(filepath) {
-        return '.' + filepath.substring(filepath.indexOf('/resources/'));
-    }
-
     function findFileName(filepath, parentDirectory) {
         return filepath.substring(filepath.indexOf('/failid/') + ('/failid/').length);
     }
-
 </script>
